% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Logistic Regression Function}
\usage{
logistic(event.e, event.c, n.e, n.c, covar)
}
\arguments{
\item{event.e}{is a vector containing the number of events in the treatment group}

\item{event.c}{is a vector containing the number of events in the control group}

\item{n.e}{is a vector containing the number of observations in the treatment group}

\item{n.c}{is a vector containing the number of observations in the control group}

\item{covar}{is a vector containing covariate values}
}
\value{
logistic meta-regression output
}
\description{
Logistic Regression Function
}
\examples{

}
